Move result to upper-level class
